+++ crates/motiongfx/Cargo.toml
@@ -12,12 +12,9 @@ repository.workspace = true
 [dependencies]
 # bevy
 bevy_ecs = { workspace = true }
-bevy_app = { workspace = true }
 bevy_platform = { workspace = true }
 bevy_math = { workspace = true }
 bevy_asset = { workspace = true, optional = true }
-bevy_transform = { workspace = true, optional = true }
-bevy_color = { workspace = true, optional = true }
 
 # other
 nonempty = { workspace = true }
@@ -26,7 +23,5 @@ nonempty = { workspace = true }
 workspace = true
 
 [features]
-default = ["asset", "transform", "color"]
+default = ["asset"]
 asset = ["dep:bevy_asset"]
-transform = ["dep:bevy_transform"]
-color = ["dep:bevy_color"]
